Implemented control to check for connectuvity when navigate to artist detail (close #23)
View variable within ConnectivityController is now public so every activity that would use it should instantiate it with a valid view object within onStart (and onResume methods when need it). This way ConnectivityController object would be able to show the snackbar to warn the user when no active Internet connection exists.